splitting up controller classes, all existing tests pass
